Updated CatFrequency
The following program enables user to call selected functions. 

Will be useful for GUI integration as it enables GUI to do the following possibility:
1. Generate the Frequency for Each Category
2. Generate Histograph for a specific category like feedback,complaints,compliment and other request

Side notes for the GUI for Histograph generation, the method requires the following inputs in order for generation to work. 
itemselect and status. 

As for getFrequency function, nothing is required to be inside.